requirements.txt
django==3.2
djangorestframework==3.12.2
fcm-django==0.3.7
opencv-python
mysqlclient
Pillow==8.4.0
djoser==2.1.0
django-webpack-loader==1.0.0
django-filter==2.4.0
django-cors-headers==3.7.0
imageio
suntime==1.2.5
django-import-export==2.7.1
uwsgi==2.0.20
django_cron==0.6.0
